backend/src/app/schemas/user.py
from pydantic import BaseModel, EmailStr


class UserCreateIn(BaseModel):
    email: EmailStr
    password: str


class UserOut(BaseModel):
    id: int
    email: str

    class Config:
        from_attributes = True
