backend/pyproject.toml
[project]
name = "pomodoro-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "prometheus-client",

  "SQLAlchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "pydantic-settings>=2.0",
]
